fix: Update input width in '_chips.scss' component
This is a minor commit which updates the width of input field from fixed
121px to 'fit-content' type, overwritten wherever the browser is able to
render this property. Also, the optional min-width and max-width are set
to limit its size between a reasnable limit of 100px and 200px. With
this change, the input placeholder (if limited to under 200px) will not
get cropped, and it will also have a better fit into the input field.

Signed-off-by: Harshit Gupta <hkg1412@gmail.com>